Complete cylinder

+++ Cube.h
@@ -89,7 +89,7 @@ private:
 
     PVList rotatePVs(Matrix mtx, PVList lst) {
         PVList res;
-        for (int i = 0; i < lst.size(); i++) {
+        for (int i = 0; i < (int)lst.size(); i++) {
             PV pv = lst[i];
             pv.p = mtx * pv.p;
             pv.v = mtx * pv.v;
